fix: Non-default branch
Fixes the case where `repositoryUrl` points to a non-default branch. Before, `checkRepository` would overwrite this value with the default branch of the repo, causing some checks to fail if the package was not being tracked in the repo's main branch. This can be the case when publishing packages off a feature/next branch before merging to main.